+++ packages/@react-native-windows/tester/src/js/components/TextInlineView.windows.js
@@ -93,7 +93,7 @@ function ClippedByText(): React.Node {
         This is an inline image
         <Image
           source={{
-            uri: 'https://picsum.photos/100',
+            uri: 'https://raw.githubusercontent.com/microsoft/react-native-windows/main/packages/@react-native/tester/js/assets/flux@3x.png',
             width: 50,
             height: 100,
           }}
@@ -131,7 +131,7 @@ class ChangeImageSize extends React.Component<mixed, ChangeSizeState> {
           This is an
           <Image
             source={{
-              uri: 'https://picsum.photos/50',
+              uri: 'https://raw.githubusercontent.com/microsoft/react-native-windows/main/packages/@react-native/tester/js/assets/flowers.png',
               width: this.state.width,
               height: 50,
             }}