+++ vnext/Folly/ThreadNameStub.cpp
@@ -2,9 +2,7 @@
 
 // Avoid bringing in a bunch of folly threading just for setThreadName
 namespace folly {
-    bool setThreadName(StringPiece)
-    {
-      return false;
-    }
-  }
-  
+bool setThreadName(StringPiece) {
+  return false;
+}
+} // namespace folly