+++ vnext/fmt/pch.h
@@ -1,2 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.