+++ packages/@react-native-windows/automation-channel/CHANGELOG.md
@@ -1,8 +1,35 @@
 # Change Log - @react-native-windows/automation-channel
 
-<!-- This log was last generated on Thu, 23 Jul 2026 23:58:53 GMT and should not be manually modified. -->
+<!-- This log was last generated on Wed, 09 Sep 2026 16:37:08 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.81.36
+
+Wed, 09 Sep 2026 16:37:08 GMT
+
+### Patches
+
+- Try fixing the e2e x86 test flakiness (vmorozov@microsoft.com)
+- Bump react-native-windows to v0.81.36
+
+## 0.81.35
+
+Sat, 22 Aug 2026 01:38:28 GMT
+
+### Patches
+
+- Make the vendored Folly headers C++23-ready and update fmt to 11.0.2 (now managed via fork-sync) (vmorozov@microsoft.com)
+- Bump react-native-windows to v0.81.35
+
+## 0.81.34
+
+Wed, 19 Aug 2026 02:25:43 GMT
+
+### Patches
+
+- Move CI pipeline to office/ISS and address security (vmorozov@microsoft.com)
+- Bump react-native-windows to v0.81.34
 
 ## 0.81.33