+++ packages/@react-native-windows/automation-commands/CHANGELOG.md
@@ -1,8 +1,33 @@
 # Change Log - @react-native-windows/automation-commands
 
-<!-- This log was last generated on Thu, 23 Jul 2026 23:58:53 GMT and should not be manually modified. -->
+<!-- This log was last generated on Wed, 09 Sep 2026 16:37:08 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.81.36
+
+Wed, 09 Sep 2026 16:37:08 GMT
+
+### Patches
+
+- Bump @react-native-windows/automation-channel to v0.81.36
+
+## 0.81.35
+
+Sat, 22 Aug 2026 01:38:28 GMT
+
+### Patches
+
+- Bump @react-native-windows/automation-channel to v0.81.35
+
+## 0.81.34
+
+Wed, 19 Aug 2026 02:25:43 GMT
+
+### Patches
+
+- Move CI pipeline to office/ISS and address security (vmorozov@microsoft.com)
+- Bump @react-native-windows/automation-channel to v0.81.34
 
 ## 0.81.33